Add Protobuf Serializer
This adds the ability to serialize `voluptuous` schemas to `.proto`
files which can be used in many other contexts such as API client generation or
docs generation.

This could be used to auto-generate home-assistant documentation
reference pages for all yaml schemas as suggested in [this post](https://github.com/home-assistant/core/issues/86050).

See the addition in the readme for an example.

With thanks to @ziv1234 who implemented serialization of dictionaries in https://github.com/home-assistant-libs/voluptuous-serialize/pull/18